fixed gui config error, removed calls to fetchProdEType()
